+++ .env.development
@@ -1,0 +1,3 @@
+NODE_ENV=development
+NEXTAUTH_SECRET=secret-key
+MONGO_URI=mongodb://soc:password@localhost:27017/
