+++ .env.production
@@ -1,1 +1,2 @@
+NEXTAUTH_URL=https://thesoc.club
 SENTRY_DSN=https://131de310ad45ea6b212e3bc6a238419d@o4506038854942720.ingest.sentry.io/4506038857564160
