+++ app/(routes)/(legal)/layout.tsx
