+++ app/(routes)/(legal)/privacy/page.tsx
