+++ app/(routes)/(legal)/styles.module.scss
