+++ app/(routes)/(legal)/terms/page.tsx
