+++ app/(routes)/about/board.ts
