+++ app/(routes)/faq/question.tsx
