+++ app/(routes)/qr/page.tsx
