+++ app/components/navbar.tsx
@@ -3,19 +3,20 @@ import Link from "next/link";
 import React from "react";
 
 import {TITLE} from "@/app/metadata";
+import styles from "./styles.module.scss";
 
-export default function Navbar(): React.ReactElement {
+export default function Navbar() {
     return (
         <header
             role="banner"
             className="flex w-full flex-row max-sm:flex-wrap items-center overflow-y-hidden mx-auto justify-between"
         >
-            <div className="flex flex-col justify-center gap-2 h-full">
+            <div className={`flex flex-col justify-center gap-2 h-full ${styles.whiteLinks}`}>
                 <h1 className="text-left text-3xl font-bold leading-tight text-white">{ TITLE }</h1>
                 <p className="text-base leading-normal mr-5 md:text-left md:text-lg">
                     The SOC Club at AUS.
                 </p>
-                <nav id="header-links" className="flex gap-2 text-white text-lg underline no-modify">
+                <nav id="header-links" className="flex gap-2 text-lg no-modify">
                     <Link href={"/"}>home</Link>
                     <Link href={"/faq"}>faq</Link>
                     <Link href={"/resources"}>resources</Link>
