+++ app/components/styles.module.scss
@@ -1,0 +1,5 @@
+.whiteLinks {
+    a[href] {
+        @apply text-white underline;
+    }
+}
